tools/RedTiger_Bot_Discord/RedTiger Bot Discord/requirements.txt
discord.py==1.6.0
datetime
requests
colorama
phonenumbers
time